requirements.txt
asgiref==3.5.2
Django==4.1.3
django-appconf==1.0.5
django-compressor==4.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-reversion==5.0.4
djangorestframework==3.14.0
pytz==2022.6
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.3
